feat: move memories
