+++ Cargo.toml
@@ -27,4 +27,4 @@ cosmian_crypto_core = { version = "10.1", default-features = false, features = [
     "sha3",
 ] }
 criterion = { version = "0.6" }
-tokio = { version = "1.45" }
+tokio = { version = "1.46", features = ["macros", "rt-multi-thread"] }
