+++ crate/findex/src/lib.rs
@@ -5,7 +5,6 @@
 // on a newer version of getrandom (0.3.2).
 #![allow(clippy::multiple_crate_versions)]
 
-mod address;
 mod adt;
 mod encoding;
 mod error;
@@ -16,8 +15,7 @@ mod ovec;
 #[cfg(any(test, feature = "test-utils"))]
 mod test_utils;
 
-pub use address::Address;
-pub use adt::{IndexADT, MemoryADT};
+pub use adt::IndexADT;
 pub use encoding::{
     Decoder, Encoder,
     generic_encoding::{generic_decode, generic_encode},
@@ -35,10 +33,3 @@ pub use encoding::{
     dummy_encoding::{WORD_LENGTH, dummy_decode, dummy_encode},
     tests::test_encoding,
 };
-
-#[cfg(any(test, feature = "test-utils"))]
-pub use memory::InMemory;
-
-/// 16-byte addresses ensure a high collision resistance that poses virtually no
-/// limitation on the index.
-pub const ADDRESS_LENGTH: usize = 16;
