+++ crate/findex/src/test_utils/benches.rs
@@ -3,11 +3,9 @@
 //! to be generic and work with any memory back end that implements the MemoryADT
 //! trait.
 
-use crate::{
-    ADDRESS_LENGTH, Address, Findex, IndexADT, MemoryADT, MemoryEncryptionLayer, WORD_LENGTH,
-    dummy_decode, dummy_encode,
-};
+use crate::{Findex, IndexADT, MemoryEncryptionLayer, WORD_LENGTH, dummy_decode, dummy_encode};
 use cosmian_crypto_core::{Secret, reexport::rand_core::CryptoRngCore};
+use cosmian_findex_memories::{ADDRESS_LENGTH, Address, MemoryADT};
 use criterion::{BenchmarkId, Criterion};
 use std::{collections::HashSet, fmt::Debug, sync::Arc};
 use tokio::runtime::{Builder, Runtime};
