+++ crate/memories/examples/example.rs
@@ -1,97 +1,97 @@
-//! This example show-cases the use of Findex to securely store a hash-map with different back ends.
-#[path = "shared_utils.rs"]
-mod shared_utils;
+// //! This example show-cases the use of Findex to securely store a hash-map with different back ends.
+// #[path = "shared_utils.rs"]
+// mod shared_utils;
 
-use cosmian_crypto_core::{CsRng, Secret, reexport::rand_core::SeedableRng};
-use cosmian_findex_memories::{
-    PostgresMemory, PostgresMemoryError, RedisMemory, SqliteMemory,
-    reexport::cosmian_findex::{ADDRESS_LENGTH, Address, Findex, IndexADT, MemoryEncryptionLayer},
-};
-use deadpool_postgres::{Config, Pool};
-use futures::executor::block_on;
-use shared_utils::{WORD_LENGTH, decoder, encoder, gen_index};
-use std::collections::HashMap;
-use tokio_postgres::NoTls;
+// use cosmian_crypto_core::{CsRng, Secret, reexport::rand_core::SeedableRng};
+// use cosmian_findex_memories::{
+//     PostgresMemory, PostgresMemoryError, RedisMemory, SqliteMemory,
+//     reexport::cosmian_findex::{ADDRESS_LENGTH, Address, Findex, IndexADT, MemoryEncryptionLayer},
+// };
+// use deadpool_postgres::{Config, Pool};
+// use futures::executor::block_on;
+// use shared_utils::{WORD_LENGTH, decoder, encoder, gen_index};
+// use std::collections::HashMap;
+// use tokio_postgres::NoTls;
 
-const DB_URL: &str = "postgres://cosmian:cosmian@localhost/cosmian";
-const DB_PATH: &str = "redis://localhost:6379";
-const DB_PATH2: &str = "./target/debug/sqlite-test.db";
-const TABLE_NAME: &str = "findex_memory";
+// const DB_URL: &str = "postgres://cosmian:cosmian@localhost/cosmian";
+// const DB_PATH: &str = "redis://localhost:6379";
+// const DB_PATH2: &str = "./target/debug/sqlite-test.db";
+// const TABLE_NAME: &str = "findex_memory";
 
-async fn create_pool(db_url: &str) -> Result<Pool, PostgresMemoryError> {
-    let mut pg_config = Config::new();
-    pg_config.url = Some(db_url.to_string());
-    let pool = pg_config.builder(NoTls)?.build()?;
-    Ok(pool)
-}
+// async fn create_pool(db_url: &str) -> Result<Pool, PostgresMemoryError> {
+//     let mut pg_config = Config::new();
+//     pg_config.url = Some(db_url.to_string());
+//     let pool = pg_config.builder(NoTls)?.build()?;
+//     Ok(pool)
+// }
 
 #[tokio::main]
 async fn main() {
-    // For cryptographic applications, it is important to use a secure RNG. In
-    // Rust, those RNG implement the `CryptoRng` trait.
-    let mut rng = CsRng::from_entropy();
+    // // For cryptographic applications, it is important to use a secure RNG. In
+    // // Rust, those RNG implement the `CryptoRng` trait.
+    // let mut rng = CsRng::from_entropy();
 
-    // Generate fresh Findex key. In practice only one user is in charge of
-    // generating the key (the administrator?): all users *must* share the same
-    // key in order to make the index inter-operable.
-    let key = Secret::random(&mut rng);
+    // // Generate fresh Findex key. In practice only one user is in charge of
+    // // generating the key (the administrator?): all users *must* share the same
+    // // key in order to make the index inter-operable.
+    // let key = Secret::random(&mut rng);
 
-    // Generating the random index.
-    let index = gen_index(&mut rng);
+    // // Generating the random index.
+    // let index = gen_index(&mut rng);
 
-    // This example uses our Redis-based implementation of `MemoryADT`.
-    let redis_memory =
-        RedisMemory::<Address<ADDRESS_LENGTH>, [u8; WORD_LENGTH]>::new_with_url(DB_PATH)
-            .await
-            .unwrap();
+    // // This example uses our Redis-based implementation of `MemoryADT`.
+    // let redis_memory =
+    //     RedisMemory::<Address<ADDRESS_LENGTH>, [u8; WORD_LENGTH]>::new_with_url(DB_PATH)
+    //         .await
+    //         .unwrap();
 
-    // You can also use our Sqlite-based implementation of `MemoryADT`.
-    let _sqlite_memory = SqliteMemory::<Address<ADDRESS_LENGTH>, [u8; WORD_LENGTH]>::new_with_path(
-        DB_PATH2,
-        TABLE_NAME.to_owned(),
-    )
-    .await
-    .unwrap();
+    // // You can also use our Sqlite-based implementation of `MemoryADT`.
+    // let _sqlite_memory = SqliteMemory::<Address<ADDRESS_LENGTH>, [u8; WORD_LENGTH]>::new_with_path(
+    //     DB_PATH2,
+    //     TABLE_NAME.to_owned(),
+    // )
+    // .await
+    // .unwrap();
 
-    // Or else, the Postgres-based implementation of `MemoryADT`. Refer to README.md for details on how to setup
-    // the database to use this example.
-    let pool = create_pool(DB_URL).await.unwrap();
-    let _postgres_memory =
-        PostgresMemory::<Address<ADDRESS_LENGTH>, [u8; WORD_LENGTH]>::new_with_pool(
-            pool.clone(),
-            TABLE_NAME.to_string(),
-        )
-        .await;
+    // // Or else, the Postgres-based implementation of `MemoryADT`. Refer to README.md for details on how to setup
+    // // the database to use this example.
+    // let pool = create_pool(DB_URL).await.unwrap();
+    // let _postgres_memory =
+    //     PostgresMemory::<Address<ADDRESS_LENGTH>, [u8; WORD_LENGTH]>::new_with_pool(
+    //         pool.clone(),
+    //         TABLE_NAME.to_string(),
+    //     )
+    //     .await;
 
-    // Adding an encryption layer to the chosen memory
-    let encrypted_memory = MemoryEncryptionLayer::new(&key, redis_memory);
+    // // Adding an encryption layer to the chosen memory
+    // let encrypted_memory = MemoryEncryptionLayer::new(&key, redis_memory);
 
-    // Instantiating Findex requires passing the key, the memory used and the
-    // encoder and decoder. Quite simple, after all :)
-    let findex = Findex::<
-        WORD_LENGTH, // size of a word
-        u64,         // type of a value
-        String,      // type of an encoding error
-        _,           // type of the memory
-    >::new(encrypted_memory, encoder, decoder);
+    // // Instantiating Findex requires passing the key, the memory used and the
+    // // encoder and decoder. Quite simple, after all :)
+    // let findex = Findex::<
+    //     WORD_LENGTH, // size of a word
+    //     u64,         // type of a value
+    //     String,      // type of an encoding error
+    //     _,           // type of the memory
+    // >::new(encrypted_memory, encoder, decoder);
 
-    // Here we insert all bindings one by one, blocking on each call. A better
-    // way would be to performed all such calls in parallel using tasks.
-    index
-        .clone()
-        .into_iter()
-        .for_each(|(kw, vs)| block_on(findex.insert(kw, vs)).expect("insert failed"));
+    // // Here we insert all bindings one by one, blocking on each call. A better
+    // // way would be to performed all such calls in parallel using tasks.
+    // index
+    //     .clone()
+    //     .into_iter()
+    //     .for_each(|(kw, vs)| block_on(findex.insert(kw, vs)).expect("insert failed"));
 
-    // In order to verify insertion was correctly performed, we search for all
-    // the indexed keywords...
-    let res = index
-        .keys()
-        .cloned()
-        .map(|kw| (kw, block_on(findex.search(&kw)).expect("search failed")))
-        .collect::<HashMap<_, _>>();
+    // // In order to verify insertion was correctly performed, we search for all
+    // // the indexed keywords...
+    // let res = index
+    //     .keys()
+    //     .cloned()
+    //     .map(|kw| (kw, block_on(findex.search(&kw)).expect("search failed")))
+    //     .collect::<HashMap<_, _>>();
 
-    // ... and verify we get the whole index back!
-    assert_eq!(res, index);
+    // // ... and verify we get the whole index back!
+    // assert_eq!(res, index);
 
     println!("All good !");
 }
