+++ crate/memories/src/databases/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(feature = "redis-mem")]
+pub(crate) mod redis_mem;
+
+#[cfg(feature = "sqlite-mem")]
+pub(crate) mod sqlite_mem;
+
+#[cfg(feature = "postgres-mem")]
+pub(crate) mod postgresql_mem;
