+++ crate/memories/src/databases/redis_mem.rs
@@ -1,6 +1,7 @@
-use cosmian_findex::{Address, MemoryADT};
 use redis::aio::ConnectionManager;
 use std::{fmt, marker::PhantomData};
+
+use crate::{MemoryADT, address::Address};
 
 // Arguments passed to the LUA script, in order:
 // 1. Guard address.
@@ -142,11 +143,13 @@ impl<const ADDRESS_LENGTH: usize, const WORD_LENGTH: usize> MemoryADT
 #[cfg(test)]
 mod tests {
 
-    use super::*;
-    use cosmian_findex::{
-        WORD_LENGTH, gen_seed, test_guarded_write_concurrent, test_rw_same_address,
-        test_single_write_and_read, test_wrong_guard,
+    use crate::test_utils::{
+        gen_seed, test_guarded_write_concurrent, test_rw_same_address, test_single_write_and_read,
+        test_wrong_guard,
     };
+
+    use super::*;
+    use cosmian_findex::WORD_LENGTH;
 
     fn get_redis_url() -> String {
         std::env::var("REDIS_HOST").map_or_else(
