+++ crate/memories/src/test_utils/mod.rs
@@ -1,5 +1,5 @@
 #[cfg(feature = "test-utils")]
-mod benches;
+mod memory_tests;
 
 #[cfg(feature = "test-utils")]
-pub use benches::*;
+pub use memory_tests::*;
